+++ examples/static.html
@@ -69,6 +69,7 @@
     <script type="importmap-shim">
       {
         "imports": {
+          "eventemitter3": "https://jspm.dev/eventemitter3",
           "three": "https://jspm.dev/three",
           "three/": "https://jspm.dev/three/"
         }
